Test: Preliminary testing of validate_timezone feat - test failure for slightly incorrect spelling
